bump pg_pinyin to 0.0.4

+++ rpmbuild/SPECS/pg_pinyin.spec
@@ -3,8 +3,12 @@
 %global sname pg_pinyin
 %global pginstdir /usr/pgsql-%{pgmajorversion}
 
+%if 0%{?pgmajorversion} < 14 || 0%{?pgmajorversion} > 18
+%{error:pg_pinyin only supports PostgreSQL 14 through 18 in PGSTY builds}
+%endif
+
 Name:		%{sname}_%{pgmajorversion}
-Version:	0.0.2
+Version:	0.0.4
 Release:	1PIGSTY%{?dist}
 Summary:	Pinyin romanization and search helpers for PostgreSQL
 License:	MIT
@@ -12,6 +16,7 @@ URL:		https://github.com/aiyou178/pg_pinyin
 Source0:	%{sname}-%{version}.tar.gz
 
 BuildRequires:	postgresql%{pgmajorversion}-devel pgdg-srpm-macros >= 1.0.27
+BuildRequires:	cargo clang rust rustfmt
 Requires:	postgresql%{pgmajorversion}-server
 
 %description
@@ -19,9 +24,25 @@ Pinyin romanization and search helpers for PostgreSQL.
 
 %prep
 %setup -q -n %{sname}-%{version}
+patch -p1 --forward -f < %{_specdir}/patches/%{sname}-%{version}-pgrx-0.18.1.patch
 
 %build
-PATH=%{pginstdir}/bin:~/.cargo/bin:$PATH cargo pgrx package -v
+export PATH=%{pginstdir}/bin:$HOME/.cargo/bin:$PATH
+
+PGRX_VERSION=0.18.1
+CURRENT_PGRX=$(cargo pgrx --version 2>/dev/null | awk '{print $2}')
+if [ "$CURRENT_PGRX" != "$PGRX_VERSION" ]; then
+	echo "cargo-pgrx $PGRX_VERSION is required; run pig build pgrx -v $PGRX_VERSION before building" >&2
+	exit 1
+fi
+cargo pgrx init --pg%{pgmajorversion}=%{pginstdir}/bin/pg_config --no-run
+CARGO_NET_GIT_FETCH_WITH_CLI=true cargo fetch
+
+# pgrx 0.18 embeds extension schema metadata in a linker section; without this
+# flag the EL9A linker can garbage-collect it and cargo-pgrx reports a missing
+# .pgrxsc section during packaging.
+export RUSTFLAGS="${RUSTFLAGS:-} -C link-arg=-Wl,--no-gc-sections"
+CARGO_NET_GIT_FETCH_WITH_CLI=true cargo pgrx package -v --no-default-features --features pg%{pgmajorversion} --pg-config %{pginstdir}/bin/pg_config
 
 %install
 rm -rf %{buildroot}
@@ -31,11 +52,17 @@ cp -a %{_builddir}/%{sname}-%{version}/target/release/%{pname}-pg%{pgmajorversio
 cp -a %{_builddir}/%{sname}-%{version}/target/release/%{pname}-pg%{pgmajorversion}/usr/pgsql-%{pgmajorversion}/share/extension/%{pname}*.sql    %{buildroot}%{pginstdir}/share/extension/
 
 %files
+%license LICENSE
+%doc readme.md README.zh-CN.md
 %{pginstdir}/lib/%{pname}.so
 %{pginstdir}/share/extension/%{pname}.control
 %{pginstdir}/share/extension/%{pname}*sql
 %exclude /usr/lib/.build-id
 
 %changelog
+* Thu Jun 11 2026 Vonng <rh@vonng.com> - 0.0.4-1PIGSTY
+- Update to upstream v0.0.4
+- Patch Cargo.toml to build with cargo-pgrx 0.18.1 for PG14-18
+
 * Wed Mar 04 2026 Vonng <rh@vonng.com> - 0.0.2-1PIGSTY
 - Initial RPM release, used by PGSTY/PIGSTY <https://pgsty.com>
